모듈화작업 ; List<Priority> 정렬기능, 출력기능, 트리 구성기능추가, Main 로직수정  - List<Priority>를 관리하는 모든 기능을 Util클래스에서 담당하도록 수정  - List<Priority> 정렬기능 : sortingList(List<Priority>)  - List<Priority> 출력기능 : printList(List<Priority>)  - 트리 구성 기능 : buildingTree(List<Priority>)  - Main로직 수정

+++ java/compression/Huffman-coding/src/com/coding/huffman/Priority.java
@@ -1,13 +1,13 @@
 package com.coding.huffman;
 
 public class Priority {
-	private int priority;
+	private float priority;
 	private char data;
 	public Priority(){}
-	public int getPriority() {
+	public float getPriority() {
 		return priority;
 	}
-	public void setPriority(int priority) {
+	public void setPriority(float priority){
 		this.priority = priority;
 	}
 	public char getData() {
